QoL
- Demande confirmation à la suppression d'un employé
- Ne demande plus confirmation quand on ferme l'application et non connecté

+++ V2/Développement/DAILY_BANK_FX_V2/src/application/view/EmployesManagementController.java
@@ -5,10 +5,12 @@ import java.util.ArrayList;
 import java.util.ResourceBundle;
 
 import application.control.EmployesManagement;
+import application.tools.AlertUtilities;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
 import javafx.scene.control.SelectionMode;
@@ -154,11 +156,14 @@ public class EmployesManagementController implements Initializable {
 
 	@FXML
 	private void doSuppEmploye() {
-		int selectedIndice = this.lvEmploye.getSelectionModel().getSelectedIndex();
-		if (selectedIndice >= 0) {
-			Employe empMod = this.ole.get(selectedIndice);
-			if (this.em.supprimerEmploye(empMod.idEmploye)) {
-				this.ole.remove(selectedIndice);
+		if (AlertUtilities.confirmYesCancel(this.primaryStage, "Suppression d'employé",
+				"Êtes vous sur de vouloir supprimer l'employé ?", null, AlertType.CONFIRMATION)) {
+			int selectedIndice = this.lvEmploye.getSelectionModel().getSelectedIndex();
+			if (selectedIndice >= 0) {
+				Employe empMod = this.ole.get(selectedIndice);
+				if (this.em.supprimerEmploye(empMod.idEmploye)) {
+					this.ole.remove(selectedIndice);
+				}
 			}
 		}
 	}
